extende the authentication package to be useful outside of Goiardi
The current implemetation of the `authentication` package makes its usage directly coupled with the `actor.Actor` type. Because of this it's not very useful as a package for use outside of the Goiardi project. I'd like to use this authentication code in my own project without having to use too many of the goiardi components.

This modification is an attempt at extending the `authentication` package without changing the API that is currently used. So this creates a new function where the authentication actually happens, `AuthenticateHeader()`, and turns `CheckHeader()` in to a wrapper function around it. Some of the references to `actor.Actor` have been replaced with just passing the public key string.