switch from ctdk/goas/v2/logger to tideland/goas/logger
The tideland/goas repo is now gone, and some of the code within `ctdk/goas` refers to it. In this PR I propose goiardi switches to the new `tideland/golib` repo.

I believe the library is mostly compatible with the changes you made to `goas/v2/logger`:

* https://github.com/tideland/golib/blob/master/logger/nosyslogger.go

Apologies for all of the import changes involved, `goimports` automatically cleaned them up.